backend/requirements.txt
Flask==2.3.2                     # Core Flask framework
mysql-connector-python==8.0.33   # MySQL database connector
Flask-Cors==3.0.10               # Enables Cross-Origin Resource Sharing (optional)
bcrypt==4.0.1                    # For secure password hashing (optional)
python-dotenv==1.0.0             # For environment variable management (optional)
Flask-Session==0.5.0
